pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
mysqlclient = "2.1.1"
pydantic = "^2.10.4"
orjson = "^3.10.12"
pytest = "^8.3.4"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
pymongo = "^4.10.1"
pytest-django = "^4.9.0"
freezegun = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
